% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcihelpr.R
\name{abci_collect_tune_res}
\alias{abci_collect_tune_res}
\title{collect tuning results from abci and transfer them to local environment}
\usage{
abci_collect_tune_res(grid_tune_id = 1001)
}
\arguments{
\item{grid_tune_id}{id for tuning with a parameter grid}
}
\description{
collect tuning results from abci and transfer them to local environment
}
