% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcihelpr.R
\name{abci_set_work_directory}
\alias{abci_set_work_directory}
\title{set work directory(.tune)  in home}
\usage{
abci_set_work_directory(abci_remote_dir = NULL, abci_local_dir = NULL)
}
\arguments{
\item{abci_remote_dir}{Remote work directory for tuning. Default is .tune}

\item{abci_local_dir}{Local work directory for tuning. Default is .tune}
}
\description{
set work directories in remote and local. if work directories do not exist, the function create them.
}
