% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcihelpr.R
\name{abci_submit_job_for_workers}
\alias{abci_submit_job_for_workers}
\title{split a job into n works and submit them to qsub (batch query)}
\usage{
abci_submit_job_for_workers(grid_tune_id = NULL, num_workers = 1)
}
\arguments{
\item{grid_tune_id}{grid_tune_id}

\item{num_workers}{number of workers. integer.}
}
\description{
split a job into n works and submit them to qsub (batch query)
}
