% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcihelpr.R
\name{abci_submit_job_to_qsub}
\alias{abci_submit_job_to_qsub}
\title{submit job to qsub query}
\usage{
abci_submit_job_to_qsub(grid_tune_id, grid_tune_start = 1, grid_tune_end = -1)
}
\arguments{
\item{grid_tune_id}{id for tuning with a parameter grid}

\item{grid_tune_start}{start no. for a subset of a parameter grid to tune}

\item{grid_tune_end}{end no. for a subset of a parameter grid to tune. -1 means the end of a parameter grid}
}
\description{
submit job to qsub query
}
