% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcihelpr.R
\name{abci_upload_params}
\alias{abci_upload_params}
\title{upload a parameter grid to abci}
\usage{
abci_upload_params(
  grid_tune_id = 1001,
  tune_wf,
  tune_folds,
  pkg_names = NULL,
  param_grid,
  fns = NULL
)
}
\arguments{
\item{grid_tune_id}{id for a parameter grid}

\item{tune_wf}{workflow object for tuning with a parameter grid}

\item{tune_folds}{resample object (rest() object) for tuning}

\item{pkg_names}{additional packages to load in the tuning script}

\item{param_grid}{A data frame of tuning combinations}

\item{fns}{character vector of function names to be exported}
}
\description{
upload a parameter grid to abci
}
