% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcihelpr.R
\name{set_abci_group_account}
\alias{set_abci_group_account}
\title{set abci group account}
\usage{
set_abci_group_account(abci_group_account)

set_abci_group_account(abci_group_account)
}
\arguments{
\item{abci_group_account}{group_account for abci}
}
\description{
set abci group account

set abci group account
}
