% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcihelpr.R
\name{set_abci_user_account}
\alias{set_abci_user_account}
\title{set abci group account}
\usage{
set_abci_user_account(abci_user_account)
}
\arguments{
\item{abci_group_account}{user_account that abci gives you}
}
\description{
set abci group account
}
