% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcihelpr.R
\name{set_mailaddr}
\alias{set_mailaddr}
\title{set mail addr to notify the end of the task}
\usage{
set_mailaddr(mailaddr)
}
\arguments{
\item{emailaddr}{mail address eg. xxxx@yyy.com}
}
\description{
set mail addr to notify the end of the task
}
