% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcihelpr.R
\name{set_ppn}
\alias{set_ppn}
\title{set the number of parallel processing per job}
\usage{
set_ppn(ppn)
}
\arguments{
\item{ppn}{number of parallel processing}
}
\description{
set the number of parallel processing per job
}
