% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcihelpr.R
\name{set_ssh_config}
\alias{set_ssh_config}
\title{change ssh config file}
\usage{
set_ssh_config(path_ssh_config, abci_remote_host)
}
\arguments{
\item{path_ssh_config}{path to ssh identity}

\item{abci_remote_host}{remote host to connect}
}
\description{
change ssh config file
}
