% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcihelpr.R
\name{set_ssh_identity}
\alias{set_ssh_identity}
\title{set ssh identify file}
\usage{
set_ssh_identity(path_ssh_identity)
}
\arguments{
\item{path_ssh_identity}{path to ssh identity}
}
\description{
set ssh identify file
}
